[mc_rbdyn][mc_tvm] Add external and compensation torques member and their equivalent joint accelerations to `Robot`
Add the new members `externalTorques_`, `compensationTorques_`,
`exteralTorquesAcc_` and `compensationTorquesAcc_` to `mc_rbdyn::Robot`.
Allong with their setters and getters `setExternalTorques`,
`setCompensationTorques`, `setExternalTorquesAcc`,
`setCompensationTorquesAcc`.
Note that the compensation torques and associated equivalent
accelerations are set as std::optional.
This is intended for later use for external forces compensation and
explicit compliance.