[mc_solver] Add feedforward compensation torques to the `DynamicConstraint` in the Tasks backend
Add an additional boolean argument to `DynamicConstraint` to enable
compensation of external torques disturbance estimation.
The boolean is defaulted to false to prevent breaking change.
The feedforward compensation torques are passed as an `Eigen::VectorXd`
to
Tasks' `MotionConstr` and `MotionSpringConstr`.
The compensation torques are obtained from
`mc_rbdyn::Robot::compensationTorques()` if available or
`mc_rbdyn::Robot::externalTorques()` otherwise.